billboards working - forgot to enable attrib arrays ><

+++ src/lib/stb_image.c
@@ -430,7 +430,7 @@ stbi_inline static uint8 get8u(stbi *s)
 static void skip(stbi *s, int n)
 {
    if (s->io.read) {
-      int blen = s->img_buffer_end - s->img_buffer;
+      int blen = ( int )( ( intptr_t )s->img_buffer_end - ( intptr_t )s->img_buffer );
       if (blen < n) {
          s->img_buffer = s->img_buffer_end;
          (s->io.skip)(s->io_user_data, n - blen);
@@ -443,7 +443,7 @@ static void skip(stbi *s, int n)
 static int getn(stbi *s, stbi_uc *buffer, int n)
 {
    if (s->io.read) {
-      int blen = s->img_buffer_end - s->img_buffer;
+      int blen = ( int )( ( intptr_t )s->img_buffer_end - ( intptr_t )s->img_buffer );
       if (blen < n) {
          int res, count;
